refactor google payments
